Initial commit of progress bar image.
The image contains two rectangles, one for showing the background of the
progress bar (with gradients green->gray depending on weekday) and one for the
actual progress bar. The idea is to clip the progress bar (id 'prog') using
javascript. Since the border of the progress bar is vital for appearance, the
size of the rectangles must be modified dynamically, since I couldn't find any
way of setting the width/height to 100% - borderwidth.